2. Тесты для идеальный вес

+++ chapter_001/src/main/java/calculator/Fit.java
@@ -2,12 +2,12 @@ package calculator;
 
 public class Fit {
 
-    public static double manWeight(short height) {
+    public static double manWeight(double height) {
         double rsl = (height - 100) * 1.15;
         return rsl;
     }
 
-    public static double womanWeight(short height) {
+    public static double womanWeight(double height) {
         double rsl = (height - 110) * 1.15;
         return rsl;
     }
